Morning commit. Bugfixes: good.price=>int32, others
UserOrderList bad

+++ PIS_Storage/PIS_Storage/DB Entities/Good.cs
@@ -19,7 +19,7 @@ namespace PIS_Storage
         public string Name { get; set; }
 
         [Required]
-        public short Price { get; set; }
+        public int Price { get; set; }
 
         [MaxLength(400)]
         public string PhotoPath { get; set; }
